(PC-20664)[API] test: adding the test for edit price or ticket only if offer more than 48h and not in status reimbursed
