+++ api/tests/core/educational/test_api.py
@@ -9,17 +9,14 @@ import pytest
 import requests_mock
 
 from pcapi.core import search
-from pcapi.core.educational import exceptions
 from pcapi.core.educational import factories as educational_factories
 from pcapi.core.educational.api import adage as educational_api_adage
 from pcapi.core.educational.api import booking as educational_api_booking
 from pcapi.core.educational.api import stock as educational_api_stock
 from pcapi.core.educational.api.offer import unindex_expired_collective_offers
-from pcapi.core.educational.models import CollectiveBooking
 from pcapi.core.educational.models import CollectiveBookingStatus
 from pcapi.core.educational.models import CollectiveStock
 import pcapi.core.offerers.factories as offerers_factories
-from pcapi.core.offers import exceptions as offers_exceptions
 import pcapi.core.search.testing as search_testing
 from pcapi.core.testing import override_settings
 import pcapi.core.users.factories as users_factories
@@ -59,336 +56,6 @@ SIMPLE_OFFER_VALIDATION_CONFIG = """
                     comparated:
                       - "suspicious"
         """
-
-
-@pytest.mark.usefixtures("db_session")
-class EditCollectiveOfferStocksTest:
-    def test_should_update_all_fields_when_all_changed(self) -> None:
-        # Given
-        initial_event_date = datetime.datetime.utcnow() + datetime.timedelta(days=5)
-        initial_booking_limit_date = datetime.datetime.utcnow() + datetime.timedelta(days=3)
-        stock_to_be_updated = educational_factories.CollectiveStockFactory(
-            beginningDatetime=initial_event_date,
-            price=1200,
-            numberOfTickets=30,
-            bookingLimitDatetime=initial_booking_limit_date,
-        )
-        new_stock_data = collective_stock_serialize.CollectiveStockEditionBodyModel(
-            beginningDatetime=datetime.datetime.now(datetime.timezone.utc)  # pylint: disable=datetime-now
-            + datetime.timedelta(days=7, hours=5),
-            bookingLimitDatetime=datetime.datetime.now(datetime.timezone.utc)  # pylint: disable=datetime-now
-            + datetime.timedelta(days=5, hours=16),
-            totalPrice=1500,
-            numberOfTickets=35,
-        )
-
-        # When
-        educational_api_stock.edit_collective_stock(
-            stock=stock_to_be_updated, stock_data=new_stock_data.dict(exclude_unset=True)
-        )
-
-        # Then
-        stock = CollectiveStock.query.filter_by(id=stock_to_be_updated.id).one()
-        assert stock.beginningDatetime == new_stock_data.beginningDatetime.replace(tzinfo=None)
-        assert stock.bookingLimitDatetime == new_stock_data.bookingLimitDatetime.replace(tzinfo=None)
-        assert stock.price == 1500
-        assert stock.numberOfTickets == 35
-
-    def test_should_update_some_fields_and_keep_non_edited_ones(self) -> None:
-        # Given
-        initial_event_date = datetime.datetime.utcnow() + datetime.timedelta(days=5)
-        initial_booking_limit_date = datetime.datetime.utcnow() + datetime.timedelta(days=3)
-        stock_to_be_updated = educational_factories.CollectiveStockFactory(
-            beginningDatetime=initial_event_date,
-            price=1200,
-            numberOfTickets=30,
-            bookingLimitDatetime=initial_booking_limit_date,
-        )
-        new_stock_data = collective_stock_serialize.CollectiveStockEditionBodyModel(
-            beginningDatetime=datetime.datetime.now(datetime.timezone.utc)  # pylint: disable=datetime-now
-            + datetime.timedelta(days=7, hours=5),
-            numberOfTickets=35,
-        )
-
-        # When
-        educational_api_stock.edit_collective_stock(
-            stock=stock_to_be_updated, stock_data=new_stock_data.dict(exclude_unset=True)
-        )
-
-        # Then
-        stock = CollectiveStock.query.filter_by(id=stock_to_be_updated.id).one()
-        assert stock.beginningDatetime == new_stock_data.beginningDatetime.replace(tzinfo=None)
-        assert stock.bookingLimitDatetime == initial_booking_limit_date
-        assert stock.price == 1200
-        assert stock.numberOfTickets == 35
-
-    def test_should_replace_bookingLimitDatetime_with_new_event_datetime_if_provided_but_none(self) -> None:
-        # Given
-        initial_event_date = datetime.datetime.utcnow() + datetime.timedelta(days=5)
-        initial_booking_limit_date = datetime.datetime.utcnow() + datetime.timedelta(days=3)
-        stock_to_be_updated = educational_factories.CollectiveStockFactory(
-            beginningDatetime=initial_event_date,
-            bookingLimitDatetime=initial_booking_limit_date,
-        )
-        new_event_datetime = datetime.datetime.now(  # pylint: disable=datetime-now
-            datetime.timezone.utc
-        ) + datetime.timedelta(days=7, hours=5)
-        new_stock_data = collective_stock_serialize.CollectiveStockEditionBodyModel(
-            beginningDatetime=new_event_datetime,
-            bookingLimitDatetime=None,
-        )
-
-        # When
-        educational_api_stock.edit_collective_stock(
-            stock=stock_to_be_updated, stock_data=new_stock_data.dict(exclude_unset=True)
-        )
-
-        # Then
-        stock = CollectiveStock.query.filter_by(id=stock_to_be_updated.id).one()
-        assert stock.bookingLimitDatetime == new_event_datetime.replace(tzinfo=None)
-
-    def test_should_replace_bookingLimitDatetime_with_old_event_datetime_if_provided_but_none_and_event_date_unchanged(
-        self,
-    ) -> None:
-        # Given
-        initial_event_date = datetime.datetime.utcnow() + datetime.timedelta(days=5)
-        initial_booking_limit_date = datetime.datetime.utcnow() + datetime.timedelta(days=3)
-        stock_to_be_updated = educational_factories.CollectiveStockFactory(
-            beginningDatetime=initial_event_date,
-            bookingLimitDatetime=initial_booking_limit_date,
-        )
-        new_stock_data = collective_stock_serialize.CollectiveStockEditionBodyModel(
-            bookingLimitDatetime=None,
-        )
-
-        # When
-        educational_api_stock.edit_collective_stock(
-            stock=stock_to_be_updated, stock_data=new_stock_data.dict(exclude_unset=True)
-        )
-
-        # Then
-        stock = CollectiveStock.query.filter_by(id=stock_to_be_updated.id).one()
-        assert stock.bookingLimitDatetime == initial_event_date
-
-    # FIXME (rpaoloni, 2022-03-09) -> None: Uncomment for when pc-13428 is merged
-    # @mock.patch("pcapi.core.search.async_index_offer_ids")
-    # def test_should_reindex_offer_on_algolia(self, mocked_async_index_offer_ids) -> None:
-    #     # Given
-    #     initial_event_date = datetime.datetime.utcnow() + datetime.timedelta(days=5)
-    #     initial_booking_limit_date = datetime.datetime.utcnow() + datetime.timedelta(days=3)
-    #     stock_to_be_updated = educational_factories.CollectiveStockFactory(
-    #         beginningDatetime=initial_event_date,
-    #         price=1200,
-    #         numberOfTickets=30,
-    #         bookingLimitDatetime=initial_booking_limit_date,
-    #     )
-    #     new_stock_data = stock_serialize.EducationalStockEditionBodyModel(
-    #         beginningDatetime=datetime.datetime.utcnow() + datetime.timedelta(days=7, hours=5),
-    #         numberOfTickets=35,
-    #     )
-
-    #     # When
-    #     edit_collective_stock(
-    #         stock=stock_to_be_updated, stock_data=new_stock_data.dict(exclude_unset=True)
-    #     )
-
-    #     # Then
-    #     stock = CollectiveStock.query.filter_by(id=stock_to_be_updated.id).one()
-    #     mocked_async_index_offer_ids.assert_called_once_with([stock.collectiveOfferId])
-
-    def test_should_not_allow_stock_edition_when_booking_status_is_not_PENDING(self) -> None:
-        # Given
-        stock_to_be_updated = educational_factories.CollectiveStockFactory(price=1200)
-        educational_factories.CollectiveBookingFactory(
-            confirmationLimitDate=datetime.datetime.utcnow() + datetime.timedelta(days=1337),
-            collectiveStock=stock_to_be_updated,
-        )
-
-        new_stock_data = collective_stock_serialize.CollectiveStockEditionBodyModel(
-            totalPrice=1500,
-        )
-
-        # When
-        with pytest.raises(exceptions.CollectiveOfferStockBookedAndBookingNotPending):
-            educational_api_stock.edit_collective_stock(
-                stock=stock_to_be_updated, stock_data=new_stock_data.dict(exclude_unset=True)
-            )
-
-        # Then
-        stock = CollectiveStock.query.filter_by(id=stock_to_be_updated.id).first()
-        assert stock.price == 1200
-
-    @freeze_time("2020-11-17 15:00:00")
-    def should_update_bookings_cancellation_limit_date_if_event_postponed(self) -> None:
-        # Given
-        educational_year = educational_factories.EducationalYearFactory(
-            beginningDate=datetime.datetime(2020, 9, 1), expirationDate=datetime.datetime(2021, 8, 31)
-        )
-        initial_event_date = datetime.datetime.utcnow() + datetime.timedelta(days=20)
-        cancellation_limit_date = datetime.datetime.utcnow() + datetime.timedelta(days=5)
-        stock_to_be_updated = educational_factories.CollectiveStockFactory(beginningDatetime=initial_event_date)
-        booking = educational_factories.CollectiveBookingFactory(
-            collectiveStock=stock_to_be_updated,
-            status=CollectiveBookingStatus.PENDING,
-            cancellationLimitDate=cancellation_limit_date,
-            confirmationLimitDate=datetime.datetime.utcnow() + datetime.timedelta(days=30),
-            educationalYear=educational_year,
-        )
-
-        new_event_date = datetime.datetime.now(  # pylint: disable=datetime-now
-            datetime.timezone.utc
-        ) + datetime.timedelta(days=25, hours=5)
-        new_stock_data = collective_stock_serialize.CollectiveStockEditionBodyModel(
-            beginningDatetime=new_event_date,
-        )
-
-        # When
-        educational_api_stock.edit_collective_stock(
-            stock=stock_to_be_updated, stock_data=new_stock_data.dict(exclude_unset=True)
-        )
-
-        # Then
-        booking_updated = CollectiveBooking.query.filter_by(id=booking.id).one()
-        assert booking_updated.cancellationLimitDate == new_event_date.replace(tzinfo=None) - datetime.timedelta(
-            days=15
-        )
-
-    @freeze_time("2020-11-17 15:00:00")
-    def should_update_bookings_cancellation_limit_date_if_beginningDatetime_earlier(self) -> None:
-        # Given
-        educational_year = educational_factories.EducationalYearFactory(
-            beginningDate=datetime.datetime(2020, 9, 1), expirationDate=datetime.datetime(2021, 8, 31)
-        )
-        initial_event_date = datetime.datetime.utcnow() + datetime.timedelta(days=20)
-        cancellation_limit_date = datetime.datetime.utcnow() + datetime.timedelta(days=5)
-        stock_to_be_updated = educational_factories.CollectiveStockFactory(beginningDatetime=initial_event_date)
-        booking = educational_factories.CollectiveBookingFactory(
-            collectiveStock=stock_to_be_updated,
-            status=CollectiveBookingStatus.PENDING,
-            cancellationLimitDate=cancellation_limit_date,
-            confirmationLimitDate=datetime.datetime.utcnow() + datetime.timedelta(days=30),
-            educationalYear=educational_year,
-        )
-
-        new_event_date = datetime.datetime.now(  # pylint: disable=datetime-now
-            datetime.timezone.utc
-        ) + datetime.timedelta(days=5, hours=5)
-        new_stock_data = collective_stock_serialize.CollectiveStockEditionBodyModel(
-            beginningDatetime=new_event_date,
-            bookingLimitDatetime=datetime.datetime.now(datetime.timezone.utc)  # pylint: disable=datetime-now
-            + datetime.timedelta(days=3, hours=5),
-        )
-
-        # When
-        educational_api_stock.edit_collective_stock(
-            stock=stock_to_be_updated, stock_data=new_stock_data.dict(exclude_unset=True)
-        )
-
-        # Then
-        booking_updated = CollectiveBooking.query.filter_by(id=booking.id).one()
-        assert booking_updated.cancellationLimitDate == datetime.datetime.utcnow()
-
-    @freeze_time("2020-11-17 15:00:00")
-    def test_should_allow_stock_edition_and_not_modify_cancellation_limit_date_when_booking_cancelled(self) -> None:
-        # Given
-        initial_event_date = datetime.datetime.utcnow() + datetime.timedelta(days=20)
-        cancellation_limit_date = datetime.datetime.utcnow() + datetime.timedelta(days=5)
-        stock_to_be_updated = educational_factories.CollectiveStockFactory(
-            beginningDatetime=initial_event_date,
-        )
-        booking = educational_factories.CollectiveBookingFactory(
-            collectiveStock=stock_to_be_updated,
-            status=CollectiveBookingStatus.CANCELLED,
-            cancellationLimitDate=cancellation_limit_date,
-            confirmationLimitDate=datetime.datetime.utcnow() + datetime.timedelta(days=30),
-        )
-
-        new_event_date = datetime.datetime.now(  # pylint: disable=datetime-now
-            datetime.timezone.utc
-        ) + datetime.timedelta(days=25, hours=5)
-        new_stock_data = collective_stock_serialize.CollectiveStockEditionBodyModel(
-            beginningDatetime=new_event_date,
-        )
-
-        # When
-        educational_api_stock.edit_collective_stock(
-            stock=stock_to_be_updated, stock_data=new_stock_data.dict(exclude_unset=True)
-        )
-
-        # Then
-        booking = CollectiveBooking.query.filter_by(id=booking.id).one()
-        assert booking.cancellationLimitDate == cancellation_limit_date
-        stock = CollectiveStock.query.filter_by(id=stock_to_be_updated.id).one()
-        assert stock.beginningDatetime == new_event_date.replace(tzinfo=None)
-
-    def test_edit_stock_of_non_approved_offer_fails(self) -> None:
-        # Given
-        offer = educational_factories.CollectiveOfferFactory(validation=OfferValidationStatus.PENDING)
-        stock_to_be_updated = educational_factories.CollectiveStockFactory(
-            collectiveOffer=offer,
-            price=1200,
-            numberOfTickets=30,
-        )
-        new_stock_data = collective_stock_serialize.CollectiveStockEditionBodyModel(
-            numberOfTickets=35,
-        )
-
-        # When
-        with pytest.raises(offers_exceptions.ApiErrors) as error:
-            educational_api_stock.edit_collective_stock(
-                stock=stock_to_be_updated, stock_data=new_stock_data.dict(exclude_unset=True)
-            )
-
-        # Then
-        assert error.value.errors == {
-            "global": ["Les offres refusées ou en attente de validation ne sont pas modifiables"]
-        }
-        stock = CollectiveStock.query.filter_by(id=stock_to_be_updated.id).one()
-        assert stock.numberOfTickets == 30
-
-    @freeze_time("2020-11-17 15:00:00")
-    def test_should_not_allow_stock_edition_when_beginningDatetime_not_provided_and_bookingLimitDatetime_set_after_existing_event_datetime(
-        self,
-    ) -> None:
-        # Given
-        stock_to_be_updated = educational_factories.CollectiveStockFactory(
-            beginningDatetime=datetime.datetime(2021, 12, 10), bookingLimitDatetime=datetime.datetime(2021, 12, 5)
-        )
-        new_stock_data = collective_stock_serialize.CollectiveStockEditionBodyModel(
-            bookingLimitDatetime=datetime.datetime(2021, 12, 20, tzinfo=datetime.timezone.utc)
-        )
-
-        # When
-        with pytest.raises(offers_exceptions.BookingLimitDatetimeTooLate):
-            educational_api_stock.edit_collective_stock(
-                stock=stock_to_be_updated, stock_data=new_stock_data.dict(exclude_unset=True)
-            )
-
-        # Then
-        stock = CollectiveStock.query.filter_by(id=stock_to_be_updated.id).one()
-        assert stock.bookingLimitDatetime == datetime.datetime(2021, 12, 5)
-
-    @freeze_time("2020-11-17 15:00:00")
-    def test_should_not_allow_stock_edition_when_bookingLimitDatetime_not_provided_and_beginningDatetime_set_before_existing_event_datetime(
-        self,
-    ) -> None:
-        # Given
-        stock_to_be_updated = educational_factories.CollectiveStockFactory(
-            beginningDatetime=datetime.datetime(2021, 12, 10), bookingLimitDatetime=datetime.datetime(2021, 12, 5)
-        )
-        new_stock_data = collective_stock_serialize.CollectiveStockEditionBodyModel(
-            beginningDatetime=datetime.datetime(2021, 12, 4, tzinfo=datetime.timezone.utc)
-        )
-
-        # When
-        with pytest.raises(offers_exceptions.BookingLimitDatetimeTooLate):
-            educational_api_stock.edit_collective_stock(
-                stock=stock_to_be_updated, stock_data=new_stock_data.dict(exclude_unset=True)
-            )
-
-        # Then
-        stock = CollectiveStock.query.filter_by(id=stock_to_be_updated.id).one()
-        assert stock.beginningDatetime == datetime.datetime(2021, 12, 10)
 
 
 @freeze_time("2020-11-17 15:00:00")
@@ -1002,84 +669,3 @@ class NotifyProUserOneDayTest:
                 }
                 assert args.kwargs["recipients"] == [booking3.collectiveStock.collectiveOffer.bookingEmails[0]]
                 assert args.kwargs["bcc_recipients"] == booking3.collectiveStock.collectiveOffer.bookingEmails[1:]
-
-    def test_edit_offer_of_cancelled_booking(self) -> None:
-        # Given
-        offer = educational_factories.CollectiveOfferFactory()
-        initial_event_date = datetime.datetime.utcnow() + datetime.timedelta(days=5)
-        initial_booking_limit_date = datetime.datetime.utcnow() + datetime.timedelta(days=3)
-        stock_to_be_updated = educational_factories.CollectiveStockFactory(
-            collectiveOffer=offer,
-            price=1200,
-            numberOfTickets=30,
-            beginningDatetime=initial_event_date,
-            bookingLimitDatetime=initial_booking_limit_date,
-        )
-        educational_factories.CollectiveBookingFactory(
-            collectiveStock=stock_to_be_updated,
-            status=CollectiveBookingStatus.CANCELLED,
-        )
-        new_stock_data = collective_stock_serialize.CollectiveStockEditionBodyModel(
-            beginningDatetime=datetime.datetime.now(datetime.timezone.utc)  # pylint: disable=datetime-now
-            + datetime.timedelta(days=7, hours=5),
-            totalPrice=1500,
-            numberOfTickets=35,
-            bookingLimitDatetime=datetime.datetime.now(datetime.timezone.utc)  # pylint: disable=datetime-now
-            + datetime.timedelta(days=5, hours=16),
-        )
-
-        # When
-        educational_api_stock.edit_collective_stock(
-            stock=stock_to_be_updated, stock_data=new_stock_data.dict(exclude_unset=True)
-        )
-
-        # Then
-        stock = CollectiveStock.query.filter_by(id=stock_to_be_updated.id).one()
-        assert stock.beginningDatetime == new_stock_data.beginningDatetime.replace(tzinfo=None)
-        assert stock.bookingLimitDatetime == new_stock_data.bookingLimitDatetime.replace(tzinfo=None)
-        assert stock.price == 1500
-        assert stock.numberOfTickets == 35
-
-    def test_edit_offer_of_other_status_booking(self) -> None:
-        # Given
-        offer = educational_factories.CollectiveOfferFactory()
-        initial_event_date = datetime.datetime.utcnow() - datetime.timedelta(days=5)
-        initial_booking_limit_date = datetime.datetime.utcnow() + datetime.timedelta(days=3)
-        educational_year = educational_factories.EducationalYearFactory(
-            beginningDate=datetime.datetime(2020, 1, 10), expirationDate=datetime.datetime(2020, 12, 12)
-        )
-        stock_to_be_updated = educational_factories.CollectiveStockFactory(
-            collectiveOffer=offer,
-            price=1200,
-            numberOfTickets=30,
-            beginningDatetime=initial_event_date,  # 2020-01-10 10:00:00
-            bookingLimitDatetime=initial_booking_limit_date,  # 2020-01-08 10:00:00
-        )
-
-        educational_factories.CollectiveBookingFactory(
-            collectiveStock=stock_to_be_updated,
-            status=CollectiveBookingStatus.CANCELLED,
-        )
-        educational_factories.CollectiveBookingFactory(
-            collectiveStock=stock_to_be_updated,
-            status=CollectiveBookingStatus.PENDING,
-            educationalYear=educational_year,
-        )
-
-        new_stock_data = collective_stock_serialize.CollectiveStockEditionBodyModel(
-            beginningDatetime=datetime.datetime.now(datetime.timezone.utc)  # pylint: disable=datetime-now
-            + datetime.timedelta(days=7, hours=5),  # 2020-01-12 15:00:00+00:00
-            totalPrice=1500,
-            numberOfTickets=35,
-            bookingLimitDatetime=datetime.datetime.now(datetime.timezone.utc)  # pylint: disable=datetime-now
-            + datetime.timedelta(days=5, hours=16),  # 2020-01-11 02:00:00+00:00
-        )
-
-        # When
-        with pytest.raises(offers_exceptions.ApiErrors) as error:
-            educational_api_stock.edit_collective_stock(
-                stock=stock_to_be_updated, stock_data=new_stock_data.dict(exclude_unset=True)
-            )
-
-        # Then
-        assert error.value.errors == {"global": ["Les évènements passés ne sont pas modifiables"]}
